Add completion rate getter + stats card list view
This commit adds a completion rate getter in the habit model, which is
now being used in the habit stats card in the statistics page as a
percentage. Additionally, this commit creates a list view in the
statistics screen which creates a habit statistics card w/ a title, and
styled grid for each habit. Added potential todos.