Add Heatmap + restructure statistics management + display
This commit adds the heatmap widget and restructures the statistics
management into a CRUD class, a summary storage class (for averages and
stats for all habits), and the habit class itself, which contains stats
for each specific habit. Finally, to better structure the data, I've
added a model that holds all statistics for a specific day.